# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Flashcards API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/answers": {
            "post": {
                "tags": [
                    "answer"
                ],
                "summary": "Create an answer",
                "operationId": "a1aa30cc6ce16b56a43129192929677d",
                "parameters": [
                    {
                        "name": "flashcardId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "text": {
                                        "type": "string"
                                    },
                                    "explanation": {
                                        "type": "string"
                                    },
                                    "is_correct": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Flashcard not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/flashcards": {
            "get": {
                "tags": [
                    "flashcard"
                ],
                "summary": "List all active flashcards",
                "description": "Return all flashcards that are not in the graveyard, for the current user, paginated",
                "operationId": "745e7a92242380cb67d3cd373765f022",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Not permitted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "flashcard"
                ],
                "summary": "Create flashcard",
                "operationId": "daad21d01dc5de91b535ea881fb18b52",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "text"
                                ],
                                "properties": {
                                    "text": {
                                        "type": "string",
                                        "example": "What colour is the sky?"
                                    },
                                    "is_true": {
                                        "type": "boolean"
                                    },
                                    "explanation": {
                                        "type": "string"
                                    },
                                    "answers": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/Answer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Not permitted"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/flashcards/all": {
            "get": {
                "tags": [
                    "flashcard"
                ],
                "summary": "List all flashcards",
                "description": "Return all flashcards - alive or buried - for the current user, paginated",
                "operationId": "75ec5dcd413c0c5b3ec8308246e6092b",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Not permitted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/flashcards/{flashcard}": {
            "get": {
                "tags": [
                    "flashcard"
                ],
                "summary": "Show a flashcard",
                "operationId": "1c30177165bc2aadc62af11696df906c",
                "parameters": [
                    {
                        "name": "flashcard",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Model not found"
                    },
                    "403": {
                        "description": "Not permitted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "flashcard"
                ],
                "summary": "Pass an answer or set of answers to the question",
                "description": "Attempt to answer the question and be judged accordingly",
                "operationId": "299249a1cadd0d7fbe7de378ddebd0de",
                "parameters": [
                    {
                        "name": "flashcard",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "answers": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Model not found"
                    },
                    "403": {
                        "description": "Not permitted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "flashcard"
                ],
                "summary": "Delete flashcard",
                "operationId": "94fe4dfd3153b01a20b068e45cad6f2c",
                "parameters": [
                    {
                        "name": "flashcard",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Model not found"
                    },
                    "403": {
                        "description": "Not permitted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "flashcard"
                ],
                "summary": "Update flashcard",
                "operationId": "edaa84d0ded0ac32545433e408066a46",
                "parameters": [
                    {
                        "name": "flashcard",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "text": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Model not found"
                    },
                    "403": {
                        "description": "Not permitted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/flashcards/graveyard": {
            "get": {
                "tags": [
                    "flashcard"
                ],
                "summary": "Get all flashcards currently in the graveyard",
                "description": "'Buried' flashcards are those which have been anwered correctly on the Hard difficulty",
                "operationId": "deb8deeaf8821d7046d97611188a5caf",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "403": {
                        "description": "Not permitted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/flashcards/random": {
            "get": {
                "tags": [
                    "flashcard"
                ],
                "summary": "Get a random flashcard",
                "description": "Gets a random flashcard, regardless of difficulty or tags",
                "operationId": "17931be0302bf83341342dfd4e3ec61b",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Model not found, OR no eligible questions"
                    },
                    "403": {
                        "description": "Not permitted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/flashcards/{flashcard}/revive": {
            "post": {
                "tags": [
                    "flashcard"
                ],
                "summary": "Resurrect a buried flashcard",
                "description": "Revive a flashcard from the graveyard back to the easy difficulty",
                "operationId": "4ad9fdbff04c4096badde7914e688b34",
                "parameters": [
                    {
                        "name": "flashcard",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Model not found"
                    },
                    "403": {
                        "description": "Not permitted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Answer": {
                "required": [
                    "text"
                ],
                "properties": {
                    "flashcard_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "text": {
                        "type": "string",
                        "example": "Blue"
                    },
                    "explanation": {
                        "type": "string",
                        "example": "Blue is the colour of the sky"
                    },
                    "is_correct": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "description": "Initial creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}